+++ edison-jobs/src/main/java/de/otto/edison/jobs/configuration/JobConfiguration.java
@@ -7,9 +7,9 @@ import de.otto.edison.jobs.repository.cleanup.StopDeadJobs;
 import de.otto.edison.jobs.repository.inmem.InMemJobRepository;
 import de.otto.edison.jobs.service.JobDefinitionService;
 import de.otto.edison.jobs.service.JobService;
+import de.otto.edison.jobs.status.JobStatusCalculator;
 import de.otto.edison.jobs.status.JobStatusDetailIndicator;
 import de.otto.edison.status.domain.Status;
-import de.otto.edison.status.domain.StatusDetail;
 import de.otto.edison.status.indicator.CompositeStatusDetailIndicator;
 import de.otto.edison.status.indicator.StatusDetailIndicator;
 import org.slf4j.Logger;
@@ -22,10 +22,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ScheduledExecutorService;
 
+import static de.otto.edison.status.domain.StatusDetail.statusDetail;
 import static java.util.concurrent.Executors.newScheduledThreadPool;
 import static java.util.stream.Collectors.toList;
 
@@ -45,9 +45,12 @@ public class JobConfiguration {
     @Value("${edison.jobs.cleanup.mark-dead-after:30}")
     int secondsToMarkJobsAsDead;
 
-    @Value("${edison.jobs.status.indicate-joberror-with-level:ERROR}")
-    String jobErrorStatusMapping;
-
+    @Value("${edison.jobs.status.calculator.default:warningOnLastJobFailed}")
+    String defaultStatusCalculator;
+    /*
+    @Value("${edison.jobs.status.calculator:}")
+    Map<String,String> statusCalculators = emptyMap();
+    */
 
     @Bean
     @ConditionalOnMissingBean(ScheduledExecutorService.class)
@@ -77,32 +80,56 @@ public class JobConfiguration {
     }
 
     @Bean
+    public JobStatusCalculator warningOnLastJobFailed(final JobRepository jobRepository) {
+        return JobStatusCalculator.warningOnLastJobFailed("warningOnLastJobFailed", jobRepository);
+    }
+
+    @Bean
+    public JobStatusCalculator errorOnLastJobFailed(final JobRepository jobRepository) {
+        return JobStatusCalculator.errorOnLastJobFailed("errorOnLastJobFailed", jobRepository);
+    }
+
+    @Bean
+    public JobStatusCalculator errorOnLastThreeJobsFailed(final JobRepository jobRepository) {
+        return JobStatusCalculator.errorOnLastNumJobsFailed("errorOnLastThreeJobsFailed", 3, jobRepository);
+    }
+
+    @Bean
+    public JobStatusCalculator errorOnLastTenJobsFailed(final JobRepository jobRepository) {
+        return JobStatusCalculator.errorOnLastNumJobsFailed("errorOnLastTenJobsFailed", 10, jobRepository);
+    }
+
+    @Bean
     @ConditionalOnProperty(name = "edison.jobs.status.enabled", havingValue = "true", matchIfMissing = true)
-    public StatusDetailIndicator jobStatusDetailIndicator(final JobDefinitionService service) {
-
-
-        final Status jobErrorMapping = parseStatusMapping(this.jobErrorStatusMapping);
-
+    public StatusDetailIndicator jobStatusDetailIndicator(final JobDefinitionService service,
+                                                          final List<JobStatusCalculator> calculators) {
         final List<JobDefinition> jobDefinitions = service.getJobDefinitions();
 
         if (jobDefinitions.isEmpty()) {
-            return () -> StatusDetail.statusDetail("Jobs", Status.OK, "No job definitions configured in application.");
+            return () -> statusDetail("Jobs", Status.OK, "No job definitions configured in application.");
         } else {
             return new CompositeStatusDetailIndicator("Jobs",
                     jobDefinitions
                             .stream()
-                            .map(d -> new JobStatusDetailIndicator(jobRepository(), d.jobName(), d.jobType(), d.maxAge(), jobErrorMapping))
+                            .map(d -> new JobStatusDetailIndicator(d, findJobStatusCalculator(d.jobType(), calculators)))
                             .collect(toList())
             );
         }
     }
 
-    Status parseStatusMapping(String statusMapping) {
-        try {
-            return Status.valueOf(statusMapping);
-        } catch (IllegalArgumentException e) {
-            String errorString = "Incorrect JobStatus Mapping (look for edison.jobs.status.indicate-joberror-with-level in your properties). Valid values are %s";
-            throw new IllegalArgumentException(String.format(errorString, Arrays.toString(Status.values())));
+    private JobStatusCalculator findJobStatusCalculator(final String jobType,
+                                                        final List<JobStatusCalculator> calculators) {
+        String key = defaultStatusCalculator;
+        /*
+        if (statusCalculators.containsKey(jobType)) {
+            key = statusCalculators.get(jobType);
         }
+        */
+        return calculators
+                .stream()
+                .filter(c -> key.equalsIgnoreCase(c.getKey()))
+                .findAny()
+                .orElseThrow(() -> new IllegalStateException("Unable to find JobStatusCalculator " + key));
     }
+
 }